Fix issue in vmbuffer wal replay
While logging visibility map changes for zheap, we don't register
the zheap buffer. That's intentional since we don't set any hint bits
in zheap buffer for tracking visibility. But, we've to store the block
number in WAL so that we can track the block number for setting the
corresponding visibility map bit while replaying the WAL for the same.