Support variable sized TPD entries.
We extend the TPD entries when (a) while reserving a slot, we found that
there aren't enough slots in TPD entry or offset-map doesn't have much
space, (b) while getting the existing TPD entry we found that offset-map
doesn't have enough space. If we find the space in the same TPD page, then
we perform inplace update of the TPD entry, otherwise, a non-inplace-update
is performed. In non-inplace-update, we mark the old entry as deleted and
later during pruning, if we encounter any deleted entry, we directly prune
it.

Currently, the implementation of in-place updates is not complete, so we
always perform non-in-place update.

Patch by Amit Kapila, Dilip Kumar and Kuntal Ghosh.